Move tests from teh Web.Tests project to the Management.Tests project, where we already have a large number of Web tests and better testing infrastructure.
