+++ Application/EdFi.Ods.AdminApp.Management.Tests/Controllers/OdsInstanceSettingsController/WhenRunningTheLearningStandardsSync.cs
@@ -1,4 +1,4 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
@@ -7,7 +7,6 @@ using System;
 using System.Configuration;
 using System.Threading.Tasks;
 using System.Web.Mvc;
-using EdFi.Ods.AdminApp.Management;
 using EdFi.Ods.AdminApp.Management.Api;
 using EdFi.Ods.AdminApp.Management.Instances;
 using EdFi.Ods.AdminApp.Web.Infrastructure.Jobs;
@@ -17,7 +16,7 @@ using Moq;
 using NUnit.Framework;
 using Shouldly;
 
-namespace EdFi.Ods.AdminApp.Web.Tests.Controllers.OdsInstanceSettingsController
+namespace EdFi.Ods.AdminApp.Management.Tests.Controllers.OdsInstanceSettingsController
 {
     [TestFixture]
     public class WhenRunningTheLearningStandardsSync
