+++ Application/EdFi.Ods.AdminApp.Management.Tests/Display/HomeScreenDisplayService/HomeScreenDisplayAssertions.cs
@@ -1,4 +1,4 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
@@ -8,7 +8,7 @@ using System.Linq;
 using EdFi.Ods.AdminApp.Web.Display.HomeScreen;
 using Shouldly;
 
-namespace EdFi.Ods.AdminApp.Web.Tests.Display.HomeScreenDisplayService
+namespace EdFi.Ods.AdminApp.Management.Tests.Display.HomeScreenDisplayService
 {
     public class HomeScreenDisplayAssertions
     {
