+++ Application/EdFi.Ods.AdminApp.Management.Tests/Display/HomeScreenDisplayService/WhenDisplayingAwsHomeScreen.cs
@@ -1,15 +1,14 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using EdFi.Ods.AdminApp.Management;
 using EdFi.Ods.AdminApp.Management.Database.Models;
 using EdFi.Ods.AdminApp.Web.Display.DisplayService;
 using EdFi.Ods.AdminApp.Web.Display.HomeScreen;
 using NUnit.Framework;
 
-namespace EdFi.Ods.AdminApp.Web.Tests.Display.HomeScreenDisplayService
+namespace EdFi.Ods.AdminApp.Management.Tests.Display.HomeScreenDisplayService
 {
     [TestFixture]
     public class WhenDisplayingAwsHomeScreen
