+++ Application/EdFi.Ods.AdminApp.Management.Tests/EdFi.Ods.AdminApp.Management.Tests.csproj
@@ -348,9 +348,24 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ActionFilters\RequireSecureConnectionFilterTests.cs" />
     <Compile Include="AssertionExtensions.cs" />
+    <Compile Include="Controllers\ControllerBaseTester.cs" />
+    <Compile Include="Controllers\EducationOrganizationsControllerTests.cs" />
+    <Compile Include="Controllers\OdsInstanceSettingsController\OdsInstanceSettingsControllerFixture.cs" />
+    <Compile Include="Controllers\OdsInstanceSettingsController\WhenGettingTheApplicationsScreenContents.cs" />
+    <Compile Include="Controllers\OdsInstanceSettingsController\WhenRunningBulkUpload.cs" />
+    <Compile Include="Controllers\OdsInstanceSettingsController\WhenRunningTheLearningStandardsSync.cs" />
+    <Compile Include="Controllers\OdsInstanceSettingsController\WhenUpdatingLearningStandards.cs" />
     <Compile Include="Database\Commands\BulkUploadValidationTests.cs" />
     <Compile Include="Database\Models\RolePermissionTests.cs" />
+    <Compile Include="Display\HomeScreenDisplayService\HomeScreenDisplayAssertions.cs" />
+    <Compile Include="Display\HomeScreenDisplayService\WhenDisplayingAwsHomeScreen.cs" />
+    <Compile Include="Display\HomeScreenDisplayService\WhenDisplayingAzureHomeScreen.cs" />
+    <Compile Include="Display\HomeScreenDisplayService\WhenDisplayingOnPremisesHomeScreen.cs" />
+    <Compile Include="Infrastructure\AutomapperTests.cs" />
+    <Compile Include="Infrastructure\CloudOdsApiConnectionInformationProviderTests\WhenGettingApiCredentials.cs" />
+    <Compile Include="Infrastructure\IO\WhenConfiguringFileImport.cs" />
     <Compile Include="Instance\GetOdsInstanceRegistrationsByUserIdQueryTests.cs" />
     <Compile Include="Instance\DeregisterOdsInstanceCommandTests.cs" />
     <Compile Include="Instance\BulkRegisterOdsInstancesRecordTests.cs" />
