+++ Application/EdFi.Ods.AdminApp.Management.Tests/Infrastructure/CloudOdsApiConnectionInformationProviderTests/WhenGettingApiCredentials.cs
@@ -1,9 +1,8 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using EdFi.Ods.AdminApp.Management;
 using EdFi.Ods.AdminApp.Management.Api;
 using EdFi.Ods.AdminApp.Web.Infrastructure;
 using Moq;
@@ -13,7 +12,7 @@ using System;
 using System.Threading.Tasks;
 using EdFi.Ods.AdminApp.Management.Instances;
 
-namespace EdFi.Ods.AdminApp.Web.Tests.Infrastructure.CloudOdsApiConnectionInformationProviderTests
+namespace EdFi.Ods.AdminApp.Management.Tests.Infrastructure.CloudOdsApiConnectionInformationProviderTests
 {
     [TestFixture]
     public class WhenGettingApiCredentials
