+++ Application/EdFi.Ods.AdminApp.Web.Tests/EdFi.Ods.AdminApp.Web.Tests.csproj
@@ -190,22 +190,7 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AutomapperTests.cs" />
-    <Compile Include="Controllers\OdsInstanceSettingsController\WhenRunningBulkUpload.cs" />
-    <Compile Include="Controllers\ControllerBaseTester.cs" />
-    <Compile Include="Controllers\OdsInstanceSettingsController\OdsInstanceSettingsControllerFixture.cs" />
-    <Compile Include="Controllers\OdsInstanceSettingsController\WhenRunningTheLearningStandardsSync.cs" />
-    <Compile Include="Controllers\OdsInstanceSettingsController\WhenUpdatingLearningStandards.cs" />
-    <Compile Include="Controllers\OdsInstanceSettingsController\WhenGettingTheApplicationsScreenContents.cs" />
-    <Compile Include="Display\HomeScreenDisplayService\HomeScreenDisplayAssertions.cs" />
-    <Compile Include="Display\HomeScreenDisplayService\WhenDisplayingAwsHomeScreen.cs" />
-    <Compile Include="Display\HomeScreenDisplayService\WhenDisplayingAzureHomeScreen.cs" />
-    <Compile Include="Display\HomeScreenDisplayService\WhenDisplayingOnPremisesHomeScreen.cs" />
-    <Compile Include="Controllers\EducationOrganizationsControllerTests.cs" />
-    <Compile Include="Infrastructure\CloudOdsApiConnectionInformationProviderTests\WhenGettingApiCredentials.cs" />
-    <Compile Include="Infrastructure\IO\WhenConfiguringFileImport.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RequireSecureConnectionFilterTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\EdFi.Ods.AdminApp.Management\EdFi.Ods.AdminApp.Management.csproj">
